+++ test/libsolidity/syntaxTests/types/function_types/definition/assign_function_via_contract_name_to_var.sol
@@ -1,0 +1,14 @@
+contract A {
+    function f() external {}
+    function g() external pure {}
+}
+
+contract B {
+    function h() external {
+        function() external f = A.f;
+        function() external pure g = A.g;
+    }
+}
+// ----
+// TypeError: (128-155): Type function A.f() is not implicitly convertible to expected type function () external.
+// TypeError: (165-197): Type function A.g() pure is not implicitly convertible to expected type function () pure external.