+++ test/libsolidity/syntaxTests/types/function_types/definition/function_via_contract_name_internal.sol
@@ -1,0 +1,11 @@
+contract A {
+    function f() internal {}
+}
+
+contract B {
+    function g() external {
+        A.f;
+    }
+}
+// ----
+// TypeError: (94-97): Member "f" not found or not visible after argument-dependent lookup in type(contract A).