+++ test/libsolidity/syntaxTests/types/function_types/definition/function_via_contract_name_overloaded.sol
@@ -1,0 +1,12 @@
+contract A {
+    function f() external {}
+    function f(uint256) external {}
+}
+
+contract B {
+    function g() external {
+        A.f;
+    }
+}
+// ----
+// TypeError: (130-133): Member "f" not unique after argument-dependent lookup in type(contract A).