+++ test/libsolidity/syntaxTests/types/function_types/definition/function_via_contract_name_public.sol
@@ -1,0 +1,9 @@
+contract A {
+    function f() public {}
+}
+
+contract B {
+    function g() external pure {
+        A.f.selector;
+    }
+}